Repository de Autor e Livro

+++ src/main/java/io/github/alexandremedeiros/libraryapi/model/Livro.java
@@ -1,5 +1,6 @@
 package io.github.alexandremedeiros.libraryapi.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Objects;
 import java.util.UUID;
@@ -37,7 +38,7 @@ public class Livro {
 	private GeneroLivro genero;
 	
 	@Column(name = "preco", precision = 18, scale = 2)
-	private Double preco;
+	private BigDecimal preco;
 	
 	@ManyToOne
 	@JoinColumn(name = "id_autor")
@@ -48,7 +49,7 @@ public class Livro {
 	}
 	
 
-	public Livro(UUID id, String isbn, String titulo, LocalDate dataPublicacao, GeneroLivro genero, Double preco,
+	public Livro(UUID id, String isbn, String titulo, LocalDate dataPublicacao, GeneroLivro genero, BigDecimal preco,
 			Autor autor) {
 		super();
 		this.id = id;
@@ -102,13 +103,16 @@ public class Livro {
 		this.genero = genero;
 	}
 
-	public Double getPreco() {
+
+	public BigDecimal getPreco() {
 		return preco;
 	}
 
-	public void setPreco(Double preco) {
+
+	public void setPreco(BigDecimal preco) {
 		this.preco = preco;
 	}
+
 
 	public Autor getAutor() {
 		return autor;
